:label: [rename] 서버 로직용 provider를 repository로 이름 변경
