+++ lib/data/remote/provider/stage_provider_impl.dart
@@ -1,18 +1,18 @@
 import 'dart:async';
 
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:pocket_pose/config/api_url.dart';
 import 'package:pocket_pose/data/entity/base_response.dart';
 import 'package:pocket_pose/data/entity/request/stage_enter_request.dart';
 import 'package:pocket_pose/data/entity/response/stage_enter_response.dart';
 import 'package:pocket_pose/data/entity/response/stage_user_list_response.dart';
+import 'package:pocket_pose/data/remote/repository/stage_repository_impl.dart';
 import 'package:pocket_pose/domain/entity/stage_talk_list_item.dart';
 import 'package:pocket_pose/domain/entity/user_list_item.dart';
 import 'package:pocket_pose/domain/provider/stage_provider.dart';
 
 class StageProviderImpl extends ChangeNotifier implements StageProvider {
+  final StageRepositoryImpl _stageRepository = StageRepositoryImpl();
+
   final List<StageTalkListItem> _talkList = [];
   final List<UserListItem> _userList = [];
   late double? _stageCurSecond;
@@ -42,86 +42,30 @@ class StageProviderImpl extends ChangeNotifier implements StageProvider {
 
   @override
   Future<BaseResponse<StageUserListResponse>> getUserList() async {
-    const storage = FlutterSecureStorage();
-    const storageKey = 'kakaoAccessToken';
-    const refreshTokenKey = 'kakaoRefreshToken';
-    String accessToken = await storage.read(key: storageKey) ?? "";
-    String refreshToken = await storage.read(key: refreshTokenKey) ?? "";
+    var response = await _stageRepository.getUserList();
 
-    var dio = Dio();
-    try {
-      dio.options.headers = {
-        "cookie": "x-access-token=$accessToken;x-refresh-token=$refreshToken"
-      };
-      dio.options.contentType = "application/json";
-      var response = await dio.get(AppUrl.stageUserListUrl);
-      var responseJson = BaseResponse<StageUserListResponse>.fromJson(
-          response.data, StageUserListResponse.fromJson(response.data['data']));
-      _userList.clear();
-      _userList.addAll(responseJson.data.list ?? []);
+    _userList.clear();
+    _userList.addAll(response.data.list ?? []);
 
-      notifyListeners();
-      return responseJson;
-    } catch (e) {
-      debugPrint("mmm StageProviderImpl catch: ${e.toString()}");
-    }
-    return throw UnimplementedError();
+    notifyListeners();
+    return response;
   }
 
   @override
   Future<BaseResponse<StageEnterResponse>> getStageEnter(
       StageEnterRequest request) async {
-    const storage = FlutterSecureStorage();
-    const storageKey = 'kakaoAccessToken';
-    const refreshTokenKey = 'kakaoRefreshToken';
-    String accessToken = await storage.read(key: storageKey) ?? "";
-    String refreshToken = await storage.read(key: refreshTokenKey) ?? "";
+    var response = await _stageRepository.getStageEnter(request);
+    _stageCurSecond = response.data.statusElapsedTime;
 
-    var dio = Dio();
-    try {
-      dio.options.headers = {
-        "cookie": "x-access-token=$accessToken;x-refresh-token=$refreshToken"
-      };
-      dio.options.contentType = "application/json";
-      var response = await dio.get(
-          '${AppUrl.stageEnterUrl}?page=${request.page}&size=${request.size}');
+    addTalkList(response.data.talkMessageData.messages ?? []);
 
-      var responseJson = BaseResponse<StageEnterResponse>.fromJson(
-          response.data, StageEnterResponse.fromJson(response.data['data']));
-      _stageCurSecond = responseJson.data.statusElapsedTime;
+    notifyListeners();
 
-      addTalkList(responseJson.data.talkMessageData.messages ?? []);
-
-      notifyListeners();
-
-      return responseJson;
-    } catch (e) {
-      debugPrint("mmm StageProviderImpl catch: ${e.toString()}");
-    }
-    throw UnimplementedError();
+    return response;
   }
 
   @override
   Future<void> getStageCatch() async {
-    const storage = FlutterSecureStorage();
-    const storageKey = 'kakaoAccessToken';
-    const refreshTokenKey = 'kakaoRefreshToken';
-    String accessToken = await storage.read(key: storageKey) ?? "";
-    String refreshToken = await storage.read(key: refreshTokenKey) ?? "";
-
-    var dio = Dio();
-    try {
-      dio.options.headers = {
-        "cookie": "x-access-token=$accessToken;x-refresh-token=$refreshToken"
-      };
-      dio.options.contentType = "application/json";
-      await dio.get(AppUrl.stageCatchUrl);
-
-      return;
-    } catch (e) {
-      debugPrint("mmm StageProviderImpl catch: ${e.toString()}");
-      if (e.toString().contains("500")) return;
-    }
-    throw UnimplementedError();
+    return await _stageRepository.getStageCatch();
   }
 }
