--- lib/domain/provider/popo_skeleton_provider.dart
@@ -1,3 +1,0 @@
-abstract class PoPoSkeletonProvider {
-  Future postSkeletonList(List<List<double>> seq);
-}
